refactor - moved and segregated the entities into packages for better readable project structure.
