+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/controllers/ConversationController.java
@@ -6,11 +6,11 @@ import dev.anirban.harmoniq_backend.dto.chat.ChatbotRequest;
 import dev.anirban.harmoniq_backend.dto.chat.ConversationDto;
 import dev.anirban.harmoniq_backend.dto.chat.ConversationRequest;
 import dev.anirban.harmoniq_backend.dto.common.ResponseWrapper;
-import dev.anirban.harmoniq_backend.entity.ChatMessage;
-import dev.anirban.harmoniq_backend.entity.Conversation;
-import dev.anirban.harmoniq_backend.service.conversation.ChatMessageService;
-import dev.anirban.harmoniq_backend.service.conversation.ChatbotService;
-import dev.anirban.harmoniq_backend.service.conversation.ConversationService;
+import dev.anirban.harmoniq_backend.entity.chat.ChatMessage;
+import dev.anirban.harmoniq_backend.entity.chat.Conversation;
+import dev.anirban.harmoniq_backend.service.chat.ChatMessageService;
+import dev.anirban.harmoniq_backend.service.chat.ChatbotService;
+import dev.anirban.harmoniq_backend.service.chat.ConversationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
