+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/controllers/ThreadController.java
@@ -6,7 +6,7 @@ import dev.anirban.harmoniq_backend.dto.thread.ThreadHistoryResponse;
 import dev.anirban.harmoniq_backend.dto.thread.ThreadRequest;
 import dev.anirban.harmoniq_backend.dto.common.ResponseWrapper;
 import dev.anirban.harmoniq_backend.dto.thread.ThreadDto;
-import dev.anirban.harmoniq_backend.entity.Thread;
+import dev.anirban.harmoniq_backend.entity.threads.Thread;
 import dev.anirban.harmoniq_backend.service.threads.ThreadService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
