+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/dto/thread/ThreadHistoryResponse.java
@@ -1,8 +1,8 @@
 package dev.anirban.harmoniq_backend.dto.thread;
 
 import dev.anirban.harmoniq_backend.dto.user.UserDto;
-import dev.anirban.harmoniq_backend.entity.Tag;
-import dev.anirban.harmoniq_backend.entity.Thread;
+import dev.anirban.harmoniq_backend.entity.threads.Tag;
+import dev.anirban.harmoniq_backend.entity.threads.Thread;
 import lombok.*;
 
 import java.util.List;
