+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/entity/threads/Like.java
@@ -1,5 +1,6 @@
-package dev.anirban.harmoniq_backend.entity;
+package dev.anirban.harmoniq_backend.entity.threads;
 
+import dev.anirban.harmoniq_backend.entity.user.User;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.UuidGenerator;
