+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/entity/threads/Thread.java
@@ -1,7 +1,8 @@
-package dev.anirban.harmoniq_backend.entity;
+package dev.anirban.harmoniq_backend.entity.threads;
 
 
 import dev.anirban.harmoniq_backend.dto.thread.ThreadDto;
+import dev.anirban.harmoniq_backend.entity.user.User;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.UuidGenerator;
