+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/repo/ChatMessageRepository.java
@@ -1,6 +1,6 @@
 package dev.anirban.harmoniq_backend.repo;
 
-import dev.anirban.harmoniq_backend.entity.ChatMessage;
+import dev.anirban.harmoniq_backend.entity.chat.ChatMessage;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
