+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/repo/CommentRepository.java
@@ -1,6 +1,6 @@
 package dev.anirban.harmoniq_backend.repo;
 
-import dev.anirban.harmoniq_backend.entity.Comment;
+import dev.anirban.harmoniq_backend.entity.threads.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, String> {
