+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/repo/InterestRepository.java
@@ -1,7 +1,7 @@
 package dev.anirban.harmoniq_backend.repo;
 
-import dev.anirban.harmoniq_backend.entity.Interest;
-import dev.anirban.harmoniq_backend.entity.User;
+import dev.anirban.harmoniq_backend.entity.user.Interest;
+import dev.anirban.harmoniq_backend.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
