+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/repo/TagRepository.java
@@ -1,6 +1,6 @@
 package dev.anirban.harmoniq_backend.repo;
 
-import dev.anirban.harmoniq_backend.entity.Tag;
+import dev.anirban.harmoniq_backend.entity.threads.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
