+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/service/auth/AuthService.java
@@ -2,7 +2,7 @@ package dev.anirban.harmoniq_backend.service.auth;
 
 import dev.anirban.harmoniq_backend.dto.auth.AuthRequest;
 import dev.anirban.harmoniq_backend.dto.auth.AuthResponse;
-import dev.anirban.harmoniq_backend.entity.User;
+import dev.anirban.harmoniq_backend.entity.user.User;
 import dev.anirban.harmoniq_backend.security.JwtService;
 import dev.anirban.harmoniq_backend.service.user.UserService;
 import lombok.RequiredArgsConstructor;
