+++ harmoniq_backend/src/main/java/dev/anirban/harmoniq_backend/service/threads/CommentService.java
@@ -1,9 +1,9 @@
 package dev.anirban.harmoniq_backend.service.threads;
 
 import dev.anirban.harmoniq_backend.dto.comment.CommentRequest;
-import dev.anirban.harmoniq_backend.entity.Comment;
-import dev.anirban.harmoniq_backend.entity.Thread;
-import dev.anirban.harmoniq_backend.entity.User;
+import dev.anirban.harmoniq_backend.entity.threads.Comment;
+import dev.anirban.harmoniq_backend.entity.threads.Thread;
+import dev.anirban.harmoniq_backend.entity.user.User;
 import dev.anirban.harmoniq_backend.repo.CommentRepository;
 import dev.anirban.harmoniq_backend.service.user.UserService;
 import lombok.RequiredArgsConstructor;
